My second project in my Data Analytics Course
In this project I'm going to make an analysis about the influenza season in the US for a fictive medical staffing agency, do determine in which states the staffing agency will need to provide temporary workers to clinics and hospitals. For this I used public data from the US Census Bureau and the Center for Disease Control and Prevention. This is a interim report which in short describes my current state in the project and displays first results of my analysis.

The company "MedCo Research" is a fictive company for which I quickly assembled a logo and a page layout, just to give it a sort of 'official' look and feel.